Refactor: move Contact UI labels to l10n extension; rename raw getter to typeLabelRaw
