+++ lib/utils/gpx_export.dart
@@ -72,7 +72,7 @@ class GpxExport {
         contact.name,
         contact.latitude!,
         contact.longitude!,
-        "Type: ${contact.typeLabel}\nPublic Key: ${contact.publicKeyHex}",
+        "Type: ${contact.typeLabelRaw}\nPublic Key: ${contact.publicKeyHex}",
         url,
       );
     }
@@ -91,7 +91,7 @@ class GpxExport {
         contact.name,
         contact.latitude!,
         contact.longitude!,
-        "Type: ${contact.typeLabel}\nPublic Key: ${contact.publicKeyHex}",
+        "Type: ${contact.typeLabelRaw}\nPublic Key: ${contact.publicKeyHex}",
         url,
       );
     }
@@ -110,7 +110,7 @@ class GpxExport {
         contact.name,
         contact.latitude ?? 0.0,
         contact.longitude ?? 0.0,
-        "Type: ${contact.typeLabel}\nPublic Key: ${contact.publicKeyHex}",
+        "Type: ${contact.typeLabelRaw}\nPublic Key: ${contact.publicKeyHex}",
         url,
       );
     }
